+++ mapbox_maps_flutter_mobile/android/src/main/kotlin/com/mapbox/maps/mapbox_maps/annotation/ControllerDelegate.kt
@@ -1,7 +1,9 @@
 package com.mapbox.maps.mapbox_maps.annotation
 
+import com.mapbox.maps.MapboxStyleManager
 import com.mapbox.maps.plugin.annotation.AnnotationManager
 
 interface ControllerDelegate {
   fun getManager(managerId: String): AnnotationManager<*, *, *, *, *, *, *>
+  fun getStyleManager(): MapboxStyleManager
 }