+++ mapbox_maps_flutter_mobile/lib/mapbox_maps_flutter_mobile.dart
@@ -29,6 +29,7 @@ part 'src/controllers/location_settings.dart';
 part 'src/controllers/style_controller.dart';
 part 'src/map_widget.dart';
 part 'src/internal/style_image_conversions.dart';
+part 'src/internal/point_annotation_icon_image.dart';
 part 'src/mapbox_map.dart';
 part 'src/mapbox_maps_options.dart';
 part 'src/mapbox_maps_platform.dart';